+++ src/MenuLayer.cpp
@@ -1,7 +1,14 @@
 #include "MenuLayer.hpp"
 #include "MenuGameLayer.hpp"
-#include "NinjaSwipeLayer.hpp"
 #include "utils/random.hpp"
+#include "utils/log.hpp"
+
+void HookedMenuLayer::onModify(auto& self) {
+    ninja::log::info("hook prio time");
+    if (!self.setHookPriority("MenuLayer::init", geode::Priority::VeryLatePost)) {
+        ninja::log::warn("failed to set hook prio??? compats may be broken");
+    }
+}
 
 bool HookedMenuLayer::init() {
     if (!MenuLayer::init()) return false;
@@ -11,14 +18,50 @@ bool HookedMenuLayer::init() {
         title->setCascadeOpacityEnabled(true);
     }
 
+    // day and night system compatibility
+    if (auto dnsEventNode = getChildByID("ninsam.day_and_night_system/Events")) {
+        ninja::log::info("day and night system momento"); 
+        dnsEventNode->setZOrder(-2);
+    }
+
+    // gd compatibility (real)
+    // ensure cascade opacity is enabled for children of buttons and such (really just circlebuttonsprite)
+    std::array<std::string, 5> menuNames = {
+        "bottom-menu",
+        "right-side-menu",
+        "top-right-menu",
+        "side-menu",
+        "close-menu"
+    };
+
+    for (auto& name : menuNames) {
+        auto menu = getChildByID(name);
+        if (!menu) continue;
+
+        for (auto& child : geode::cocos::CCArrayExt<cocos2d::CCNode*>(menu->getChildren())) {
+            for (auto& innerChild : geode::cocos::CCArrayExt<cocos2d::CCNode*>(child->getChildren())) {
+                auto castInnerChild = geode::cast::typeinfo_cast<cocos2d::CCNodeRGBA*>(innerChild);
+                if (!castInnerChild) continue;
+                castInnerChild->setCascadeOpacityEnabled(true);
+            }
+        }
+    }
+
+
+    // add our very own brand spanking new shiny icon ninja custom layer!!!!!!!!
+    m_fields->ninjaSwipeLayer = NinjaSwipeLayer::create();
+    addChild(m_fields->ninjaSwipeLayer, -1);
+    
+    m_menuGameLayer->setZOrder(-5); // get back there
+
     return true;
 }
 
 void HookedMenuLayer::keyDown(cocos2d::enumKeyCodes code) {
-    auto mgl = static_cast<HookedMenuGameLayer*>(m_menuGameLayer);
-    auto nsl = mgl->m_fields->ninjaSwipeLayer;
-    if (nsl->m_state == State::Gameplay) {
-        if (code == cocos2d::enumKeyCodes::KEY_Escape) nsl->exitGameplay(nullptr);
+    auto ninja = m_fields->ninjaSwipeLayer;
+
+    if (ninja->m_state == State::Gameplay) {
+        if (code == cocos2d::enumKeyCodes::KEY_Escape) ninja->exitGameplay(nullptr);
         // dont call orig
         return;
     }
@@ -35,10 +78,10 @@ void HookedMenuLayer::runEnterGameplayAnimations() {
 
     // dw about this code
     // rewrite it if you wish
-    // gets all children and if they're not this figures out where they should
-    // move to and stores it in (std::map)m_fields->movedNodeMovements
+    // gets all children and if they should be moved figures out where they
+    // should move to and stores it in (std::map)m_fields->movedNodeMovements
     for (auto& child : geode::cocos::CCArrayExt<CCNode*>(getChildren())) {
-        if (child == m_menuGameLayer) continue;
+        if (child == m_menuGameLayer || child == m_fields->ninjaSwipeLayer) continue;
         
         // get side + move dist
         cocos2d::CCPoint movement;